dkilcy/openstack/pip_requirements.txt
#  pip install -r pip_requirements.txt

python-novaclient
python-ceilometerclient

